chore: split optimization test, useCommitCount
