+++ tests/optimization.test.tsx
@@ -1,0 +1,62 @@
+import React, { Fragment, useRef } from 'react'
+import { fireEvent, render, waitFor } from '@testing-library/react'
+import { Provider as ProviderOrig, atom, useAtom } from '../src/index'
+
+const Provider = process.env.PROVIDER_LESS_MODE ? Fragment : ProviderOrig
+
+it('only relevant render function called (#156)', async () => {
+  const count1Atom = atom(0)
+  const count2Atom = atom(0)
+
+  const Counter1: React.FC = () => {
+    const [count, setCount] = useAtom(count1Atom)
+    const renderCount = useRef(0)
+    ++renderCount.current
+    return (
+      <>
+        <div>
+          count1: {count} ({renderCount.current})
+        </div>
+        <button onClick={() => setCount((c) => c + 1)}>button1</button>
+      </>
+    )
+  }
+
+  const Counter2: React.FC = () => {
+    const [count, setCount] = useAtom(count2Atom)
+    const renderCount = useRef(0)
+    ++renderCount.current
+    return (
+      <>
+        <div>
+          count2: {count} ({renderCount.current})
+        </div>
+        <button onClick={() => setCount((c) => c + 1)}>button2</button>
+      </>
+    )
+  }
+
+  const { getByText } = render(
+    <Provider>
+      <Counter1 />
+      <Counter2 />
+    </Provider>
+  )
+
+  await waitFor(() => {
+    getByText('count1: 0 (1)')
+    getByText('count2: 0 (1)')
+  })
+
+  fireEvent.click(getByText('button1'))
+  await waitFor(() => {
+    getByText('count1: 1 (2)')
+    getByText('count2: 0 (1)')
+  })
+
+  fireEvent.click(getByText('button2'))
+  await waitFor(() => {
+    getByText('count1: 1 (2)')
+    getByText('count2: 1 (2)')
+  })
+})
